Fixed WebPack issue.. Moved reducer to store.
