+++ Assignments/Assignment5/src/index.js
@@ -1,16 +1,14 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import { createStore } from 'redux';
+import { Provider } from 'react-redux';
+
 import './index.css';
 import App from './App';
 import registerServiceWorker from './registerServiceWorker';
-
-import { createStore } from 'redux';
-import { Provider } from 'react-redux';
-import reducer from './reducer';
+import reducer from './store/reducer';
 
 const store = createStore(reducer);
 
-ReactDOM.render(<Provider store={store}> <App /> </Provider>, document.getElementById('root'));
-//ReactDOM.render(<App />, document.getElementById('root'));
-
+ReactDOM.render(<Provider store={store}><App /></Provider>, document.getElementById('root'));
 registerServiceWorker();
